requirements.txt
solara
youtube-transcript-api
langchain
langchain-community
faiss-cpu
google-generativeai
python-dotenv